chore(cli): improve tooling around deposits
